Bringing in Rcpp(Eigen); reapplying previously applied tweaks

+++ src/reg-lib/_reg_f3d.cpp
@@ -242,7 +242,7 @@ void reg_f3d<T>::Initialise()
       if(this->spacing[2]!=this->spacing[2]) this->spacing[2]=this->spacing[0];
 
       /* Convert the spacing from voxel to mm if necessary */
-      float spacingInMillimeter[3]= {this->spacing[0],this->spacing[1],this->spacing[2]};
+      float spacingInMillimeter[3]= {static_cast<float>(this->spacing[0]),static_cast<float>(this->spacing[1]),static_cast<float>(this->spacing[2])};
       if(this->usePyramid)
       {
          if(spacingInMillimeter[0]<0) spacingInMillimeter[0] *= -1.0f * this->referencePyramid[this->levelToPerform-1]->dx;
